uyga vazifa 16kun c lang

+++ 02day/tubmi.c
@@ -5,7 +5,7 @@ int sum(int n){
 	if (n == 1){
 		return 0;
 	}
-	return n+ n % sum(n-1);
+	return n+ (int)n % sum(n-1);
 }
 
 int main(){
